:truck: Refactor #21: 스페이스 화면 - 라벨 탭 관련 ui 계층 폴더 이동
