+++ app/src/main/java/com/umc/edison/presentation/space/LabelListViewModel.kt
@@ -8,7 +8,7 @@ import com.umc.edison.domain.usecase.label.UpdateLabelUseCase
 import com.umc.edison.presentation.base.BaseViewModel
 import com.umc.edison.presentation.model.LabelModel
 import com.umc.edison.presentation.model.toPresentation
-import com.umc.edison.ui.space.EditMode
+import com.umc.edison.ui.label.EditMode
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
