+++ app/src/main/java/com/umc/edison/ui/label/LabelDetailScreen.kt
@@ -1,4 +1,4 @@
-package com.umc.edison.ui.space
+package com.umc.edison.ui.label
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
