+++ app/src/main/java/com/umc/edison/ui/label/LabelEditModal.kt
@@ -1,4 +1,4 @@
-package com.umc.edison.ui.space
+package com.umc.edison.ui.label
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
